*** empty log message ***

+++ v2/ulib_std.cpp
@@ -265,7 +265,7 @@ double mtof(double f)
 {
     if (f <= -1500) return(0);
     else if (f > 1499) return(mtof(1499));
-    else return (float)(8.17579891564 * exp(.0577622650 * (double)f));
+    else return (8.17579891564 * exp(.0577622650 * f));
 }
 
 CK_DLL_SFUN( mtof_impl )
